refactor: move DatePicker, InformationInput, and Selector to components/forms/tools directory
